refactor: move requests to domain package
