+++ src/main/kotlin/se/uulm/snowballr/backend/model/incoming/criterion/UpdateCriterionRequest.kt
@@ -1,12 +1,12 @@
-package se.uulm.snowballr.backend.model.incoming
+package se.uulm.snowballr.backend.model.incoming.criterion
 
 import se.uulm.snowballr.backend.model.dto.criterion.CriterionCategory
 import java.util.UUID
 
-data class CreateCriterionRequest(
+data class UpdateCriterionRequest(
+    val criterionId: UUID,
     val tag: String,
     val name: String,
     val description: String,
     val category: CriterionCategory,
-    val projectId: UUID?,
 )
